sw/requirements.txt
colorama==0.4.6
docopt==0.6.2
pyserial==3.5
